add PyrenewWastewaterData
