+++ pyrenew_hew/pyrenew_wastewater_data.py
@@ -1,0 +1,177 @@
+import datetime
+from typing import Self
+
+import jax
+import jax.numpy as jnp
+import polars as pl
+from jax.typing import ArrayLike
+
+
+class PyrenewWastewaterData:
+    """
+    Class for holding wastewater input data
+    to a PyrenewHEW model.
+    """
+
+    def __init__(
+        self,
+        data_observed_disease_wastewater: pl.DataFrame = None,
+        population_size: int = None,
+        pop_fraction: ArrayLike = jnp.array([1]),
+    ) -> None:
+        self.data_observed_disease_wastewater = (
+            data_observed_disease_wastewater
+        )
+        self.population_size = population_size
+        self.pop_fraction = pop_fraction
+
+    @property
+    def site_subpop_spine(self):
+        ww_data_present = self.data_observed_disease_wastewater is not None
+        if ww_data_present:
+            add_auxiliary_subpop = (
+                self.population_size
+                > self.data_observed_disease_wastewater.select(
+                    pl.col("site_pop", "site", "lab", "lab_site_index")
+                )
+                .unique()
+                .get_column("site_pop")
+                .sum()
+            )
+            site_indices = (
+                self.data_observed_disease_wastewater.select(
+                    ["site_index", "site", "site_pop"]
+                )
+                .unique()
+                .sort("site_index")
+            )
+            if add_auxiliary_subpop:
+                aux_subpop = pl.DataFrame(
+                    {
+                        "site_index": [None],
+                        "site": [None],
+                        "site_pop": [
+                            self.population_size
+                            - site_indices.select(pl.col("site_pop"))
+                            .get_column("site_pop")
+                            .sum()
+                        ],
+                    }
+                )
+            else:
+                aux_subpop = pl.DataFrame()
+            site_subpop_spine = (
+                pl.concat([aux_subpop, site_indices], how="vertical_relaxed")
+                .with_columns(
+                    subpop_index=pl.col("site_index")
+                    .cum_count()
+                    .alias("subpop_index"),
+                    subpop_name=pl.format(
+                        "Site: {}", pl.col("site")
+                    ).fill_null("remainder of population"),
+                )
+                .rename({"site_pop": "subpop_pop"})
+            )
+        else:
+            site_subpop_spine = pl.DataFrame(
+                {
+                    "site_index": [None],
+                    "site": [None],
+                    "subpop_pop": [self.population_size],
+                    "subpop_index": [1],
+                    "subpop_name": ["total population"],
+                }
+            )
+        return site_subpop_spine
+
+    @property
+    def date_time_spine(self):
+        if self.data_observed_disease_wastewater is not None:
+            date_time_spine = pl.DataFrame(
+                {
+                    "date": pl.date_range(
+                        start=self.date_observed_disease_wastewater.min(),
+                        end=self.date_observed_disease_wastewater.max(),
+                        interval="1d",
+                        eager=True,
+                    )
+                }
+            ).with_row_index("t")
+            return date_time_spine
+
+    @property
+    def wastewater_data_extended(self):
+        if self.data_observed_disease_wastewater is not None:
+            return (
+                self.data_observed_disease_wastewater.join(
+                    self.date_time_spine, on="date", how="left", coalesce=True
+                )
+                .join(
+                    self.site_subpop_spine,
+                    on=["site_index", "site"],
+                    how="left",
+                    coalesce=True,
+                )
+                .with_row_index("ind_rel_to_observed_times")
+            )
+
+    @property
+    def date_observed_disease_wastewater(self):
+        if self.data_observed_disease_wastewater is not None:
+            return self.data_observed_disease_wastewater["date"].to_numpy()
+
+    @property
+    def data_observed_disease_wastewater_conc(self):
+        if self.data_observed_disease_wastewater is not None:
+            return self.wastewater_data_extended[
+                "log_genome_copies_per_ml"
+            ].to_numpy()
+
+    @property
+    def ww_censored(self):
+        if self.data_observed_disease_wastewater is not None:
+            return self.wastewater_data_extended.filter(
+                pl.col("below_lod") == 1
+            )["ind_rel_to_observed_times"].to_numpy()
+        return None
+
+    @property
+    def ww_uncensored(self):
+        if self.data_observed_disease_wastewater is not None:
+            return self.wastewater_data_extended.filter(
+                pl.col("below_lod") == 0
+            )["ind_rel_to_observed_times"].to_numpy()
+
+    @property
+    def ww_observed_times(self):
+        if self.data_observed_disease_wastewater is not None:
+            return self.wastewater_data_extended["t"].to_numpy()
+
+    @property
+    def ww_observed_subpops(self):
+        if self.data_observed_disease_wastewater is not None:
+            return self.wastewater_data_extended["subpop_index"].to_numpy()
+
+    @property
+    def ww_observed_lab_sites(self):
+        if self.data_observed_disease_wastewater is not None:
+            return self.wastewater_data_extended["lab_site_index"].to_numpy()
+
+    @property
+    def ww_log_lod(self):
+        if self.data_observed_disease_wastewater is not None:
+            return self.wastewater_data_extended["log_lod"].to_numpy()
+
+    @property
+    def n_ww_lab_sites(self):
+        if self.data_observed_disease_wastewater is not None:
+            return self.wastewater_data_extended["lab_site_index"].n_unique()
+
+    @property
+    def lab_site_to_subpop_map(self):
+        if self.data_observed_disease_wastewater is not None:
+            return (
+                self.wastewater_data_extended["lab_site_index", "subpop_index"]
+                .unique()
+                .sort(by="lab_site_index")
+            )["subpop_index"].to_numpy()
